# Optimise packetio.Buffer by using a circular buffer