This more than doubles the performance of the current (highly unrealistic) benchmark, and reduces allocation to virtually nothing.  There are some minor API changes, described in the commit messages.

```
name          old time/op    new time/op    delta
Buffer14-4       552ns ± 8%     176ns ± 2%   -68.10%  (p=0.000 n=8+8)
Buffer140-4      435ns ± 5%     227ns ± 2%   -47.81%  (p=0.000 n=8+8)
Buffer1400-4     929ns ± 3%     385ns ± 2%   -58.55%  (p=0.000 n=8+8)

name          old speed      new speed      delta
Buffer14-4    25.4MB/s ± 8%  79.5MB/s ± 2%  +212.67%  (p=0.000 n=8+8)
Buffer140-4    322MB/s ± 5%   617MB/s ± 2%   +91.49%  (p=0.000 n=8+8)
Buffer1400-4  1.51GB/s ± 3%  3.64GB/s ± 2%  +141.27%  (p=0.000 n=8+8)

name          old alloc/op   new alloc/op   delta
Buffer14-4        138B ± 2%        3B ± 0%   -97.83%  (p=0.000 n=6+8)
Buffer140-4       264B ± 1%        4B ± 0%   -98.48%  (p=0.000 n=7+8)
Buffer1400-4    1.45kB ± 0%    0.01kB ± 0%   -99.52%  (p=0.000 n=8+8)

name          old allocs/op  new allocs/op  delta
Buffer14-4        1.00 ± 0%      0.00       -100.00%  (p=0.000 n=8+8)
Buffer140-4       1.00 ± 0%      0.00       -100.00%  (p=0.000 n=8+8)
Buffer1400-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=8+8)
```