@Sean-Der 2020-06-02
absolutely amazing, nice work @jech 

Mind adding a test that does [AllocsPerRun](https://golang.org/pkg/testing/#AllocsPerRun) just to make sure someone else doesn't regress your work.

After that I can squash and merge!