@jech 2020-06-02
Done as you requested.  While I was at it, I added some logic to prevent the allocation from growing beyond what was specified by `SetLimitSize`.