@Sean-Der 2020-06-05
@jech @at-wat I vote we merge this one. No breaking/behavior changes, and covered with tests!

I am all for simplifying in the future. If you want to cut TODOs and assign them to me and @jech @at-wat I want to make sure we all are in this together :) I just don't want to lose a contributor if we move too slow :(