@jech 2020-06-05
> I'm currently don't have actual application that exceeds 4MB limit.

Please feel free to rework the code to remove the limit, but please do it in a separate commit rather than editing mine.

> But the size of the streaming contents today is increasing, this change prohibit the package to be use in such applications.

Please be aware that the data structure I'm using is not necessarily adapted to very large buffers (on the orders of gigabytes or so).  If you need such huge buffers, you might need to rework the data structurs.