@at-wat 2020-07-05
Sorry, I had merged this without checking lint error.
Edited and force pushed two commit messages to fix lint error on master branch.